% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.nwqs}
\alias{coef.nwqs}
\title{Extract Coefficients from NWQS Objects}
\usage{
\method{coef}{nwqs}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"nwqs"}.}

\item{...}{Additional arguments.}
}
\value{
A named numeric vector of averaged global coefficients.
}
\description{
Extract Coefficients from NWQS Objects
}
