% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.nwqs}
\alias{plot.nwqs}
\title{Plot Diagnostics for NWQS}
\usage{
\method{plot}{nwqs}(
  x,
  type = c("both", "curves", "weights"),
  y_scale = c("partial", "predicted"),
  components = NULL,
  overlay = TRUE,
  plot_ci = TRUE,
  base_size = 14,
  ...
)
}
\arguments{
\item{x}{An object of class "nwqs".}

\item{type}{"both", "curves", 或 "weights".}

\item{y_scale}{"partial" (从0开始的偏效应) 或 "predicted" (加上截距的绝对预测值，让曲线悬浮).}

\item{components}{字符向量。指定要画的污染物名称。}

\item{overlay}{逻辑值。TRUE表示叠加，FALSE分面。}

\item{plot_ci}{逻辑值。是否绘制置信区间。}

\item{base_size}{Integer. Base font size for ggplot2.}

\item{...}{额外参数。}
}
\description{
Plot Diagnostics for NWQS
}
