% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwqs.R
\name{plot_nwqs_contrast_box}
\alias{plot_nwqs_contrast_box}
\title{Plot Grouped Boxplot for NWQS Quantile Contrasts
绘制 NWQS 联合与各个成分暴露分位数对比的分组箱线图 (常规原始尺度)}
\usage{
plot_nwqs_contrast_box(model, exponentiate = NULL)
}
\arguments{
\item{model}{nwqs 模型结果对象 (必须是 rh > 1 的结果)}

\item{exponentiate}{logical. 是否转换为 OR 值 (默认为 NULL, 若为 binomial 自动转换)}
}
\description{
Plot Grouped Boxplot for NWQS Quantile Contrasts
绘制 NWQS 联合与各个成分暴露分位数对比的分组箱线图 (常规原始尺度)
}
