% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwqs.R
\name{calc_weight_error}
\alias{calc_weight_error}
\title{Calculate Weight Allocation Error Metrics
计算权重分配的 SAE 和 RMSE}
\usage{
calc_weight_error(w_est, w_true)
}
\arguments{
\item{w_est}{numeric vector. 模型估计出的权重 (Estimated weights).}

\item{w_true}{numeric vector. 真实的设定权重 (True weights).}
}
\value{
A list containing SAE and RMSE.
}
\description{
Calculate Weight Allocation Error Metrics
计算权重分配的 SAE 和 RMSE
}
