% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.nwqs_boot}
\alias{coef.nwqs_boot}
\title{Extract Coefficients from nwqs_boot Objects}
\usage{
\method{coef}{nwqs_boot}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"nwqs_boot"}.}

\item{...}{Additional arguments.}
}
\value{
A named numeric vector of point-estimate coefficients.
}
\description{
Extract Coefficients from nwqs_boot Objects
}
