% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_nwqs_effects}
\alias{extract_nwqs_effects}
\title{Extract Effects and Empirical Bootstrap CI from NWQS Object (真·全局经验分位数版)}
\usage{
extract_nwqs_effects(model_res, return_raw = FALSE)
}
\description{
Extract Effects and Empirical Bootstrap CI from NWQS Object (真·全局经验分位数版)
}
