% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{generate_covariates}
\alias{generate_covariates}
\title{生成流行病学测量的混杂因素集 (Generate Covariates)}
\usage{
generate_covariates(
  n_obs = 1000,
  beta_cont = 0.5,
  beta_bin = -0.8,
  beta_cat = c(0, -0.5, 0.7),
  prob_bin = 0.5,
  prob_cat = c(1/3, 1/3, 1/3),
  Intercept = 0
)
}
\arguments{
\item{n_obs}{Integer。模拟队列的样本量。}

\item{beta_cont}{Numeric。连续型变量（如年龄的标准化值）的真实回归系数。}

\item{beta_bin}{Numeric。二分类变量（如性别）的真实回归系数。}

\item{beta_cat}{Numeric vector。多分类变量（如吸烟状态类别）的真实回归系数向量。}

\item{prob_bin}{Numeric。二分类变量发生率。}

\item{prob_cat}{Numeric vector。多分类变量各水平的分布概率。}

\item{Intercept}{Numeric。基线截距。}
}
\value{
List。包含生成的协变量数据框 \code{original}、设计矩阵形式的 \code{mm}，以及真实的线性预测子贡献 \code{eta_cov}。
}
\description{
随机生成包含连续型、二分类和多分类变量的混杂因素数据集，并计算它们的真实线性效应。
}
\details{
此函数用于模拟流行病学研究中常规收集的、需要被模型控制的协变量（如年龄、性别、BMI 分组或吸烟状态）。
将这些协变量注入生成数据，旨在验证模型在存在可测量混杂 (Measured Confounding) 时，
对混合物主效应进行无偏估计 (Unbiased Estimation) 的能力。
}
