% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{generate_linear_data}
\alias{generate_linear_data}
\title{生成线性混合物效应的连续型数据 (Generate Linear Model Data)}
\usage{
generate_linear_data(
  n_obs = 1000,
  mu_preds,
  sigma_preds,
  beta_wqs = 1,
  beta_preds,
  snr_db = 10,
  transform_fun = NULL,
  seed = NULL,
  ...
)
}
\description{
为 Monte Carlo 模拟生成标准线性响应数据。暴露组分通过多元正态分布生成，
可选经过分位数转换，最后与协变量的线性预测子组合，并根据信噪比引入测量误差。
}
